requirements.txt
fastapi
uvicorn
openai
faiss-cpu
python-docx
python-pptx
pydantic
langchain
